Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -3,12 +3,12 @@
 
 
 int main() {
-    int a,b,c;
+    float a,b,c;
     float avg;
     scanf("%d",&a);
     scanf("%d",&b);
     scanf("%d",&c);
-    avg = (float(a+b+c))/3;
+    avg = (a+b+c)/3;
     printf("Average: %.2f",avg);
     
 }
